+++ src/OHardlyWorkingBot/BotDestabilizer.java
@@ -1,0 +1,13 @@
+package OHardlyWorkingBot;
+
+import battlecode.common.*;
+
+public class BotDestabilizer extends Utils{
+    public static void initDestabilizer() throws GameActionException{
+
+    }
+
+    public static void runDestabilizer() throws GameActionException{
+        
+    }
+}